Phase 1 - Basic BNF Parser Generator
* Recursive Descent Parser Library
* BNF Grammar Parser based on SimpleParse
* In-code Parser Generator

Parsers can be generated from an input BNF file.  No preprocessor
support at this time; the subsystem may be broken out into a distinct
project for that.  Both the BNF parser and the generated parser use the
same parsing framework.  The output of the generated parser is an AST
suitable for most grammars.

At the moment the implementation does not feature a distinct lexer, so
whitespace tokens are generated and placed in the AST unless they are
explicitly labeled as 'unreported'.